Deleted unnessecary autos added constants sorted paths.
