+++ src/main/java/frc/robot/Constants.java
@@ -77,4 +77,9 @@ public final class Constants {
 
         public static Rotation2d forwardAngle = new Rotation2d(Math.toRadians(0));
     }
+        public static final class Paths {
+            public static Translation2d START_LEFT = new Translation2d(0,7.054);
+            public static Translation2d START_CENTER = new Translation2d(0,0); //Find Real Value
+            public static Translation2d START_RIGHT = new Translation2d(0,1.622);
+        }
 }
