+++ src/main/java/frc/robot/auto/paths/centerStart/twoPiece/CenterNearRight.java
@@ -2,11 +2,12 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.auto.paths;
+package frc.robot.auto.paths.centerStart.twoPiece;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj2.command.PrintCommand;
+import frc.robot.Constants;
 import frc.robot.auto.pathing.pathObjects.Path;
 import frc.robot.auto.pathing.pathObjects.PathPoint;
 
@@ -14,14 +15,14 @@ import frc.robot.auto.pathing.pathObjects.PathPoint;
  * Example pure pursuit path, showing a basic
  * and readable implementation.
  */
-public class LeftNearLeftNearCenter extends Path {
+public class CenterNearRight extends Path {
     /**
      * Drives the robot 1.5 meters along the x
      * axis (which should be forward, if the settings
      * are properly configured within the PathingConstants
      * class), at the robots maximum speed.
      */
-    public LeftNearLeftNearCenter() {
+    public CenterNearRight() {
         /**
          * Initializes Path super with the given
          * set of points, the first point passed
@@ -29,22 +30,16 @@ public class LeftNearLeftNearCenter extends Path {
          */
         super(            
             new PathPoint(
-                new Translation2d(0,7.054),               // Position (meters)
+                Constants.Paths.START_CENTER,               // Position (meters)
                 new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
                 1,    // Speed (m/s)
                 new PrintCommand("Shoot")       // Command 
             ),
             new PathPoint(
-                new Translation2d(2.757,7.054),               // Position (meters)
+                new Translation2d(3.348,4.054),               // Position (meters)
                 new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
                 1,    // Speed (m/s)
                 new PrintCommand("Shoot")       // Command 
-            ),
-            new PathPoint(
-                new Translation2d(3.348,5.676),               // Position (meters)
-                new Rotation2d(Math.toRadians(-Math.PI)),     // Rotation (rad)
-                1,    // Speed (m/s)
-                new PrintCommand("Shoot")      // Command 
             )
         );
     }
