+++ src/main/java/frc/robot/auto/paths/leftStart/threePiece/LeftNearLeftMidLeft.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.auto.paths;
+package frc.robot.auto.paths.leftStart.threePiece;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
