+++ src/main/java/frc/robot/auto/paths/leftStart/twoPiece/LeftNearLeft.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.auto.paths;
+package frc.robot.auto.paths.leftStart.twoPiece;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
@@ -14,44 +14,31 @@ import frc.robot.auto.pathing.pathObjects.PathPoint;
  * Example pure pursuit path, showing a basic
  * and readable implementation.
  */
-public class CenterNearCenter extends Path {
+public class LeftNearLeft extends Path {
     /**
      * Drives the robot 1.5 meters along the x
      * axis (which should be forward, if the settings
      * are properly configured within the PathingConstants
      * class), at the robots maximum speed.
      */
-    public CenterNearCenter() {
+    public LeftNearLeft() {
         /**
          * Initializes Path super with the given
          * set of points, the first point passed
          * into super is the first point along the path.
          */
         super(            
-            0.5, // End tolerance
             new PathPoint(
-                new Translation2d(0,0),               // Position (meters)
-                new Rotation2d(Math.toRadians(180)),     // Rotation (rad)
-                3,    // Speed (m/s)
+                new Translation2d(0,7.054),               // Position (meters)
+                new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
+                1,    // Speed (m/s)
                 new PrintCommand("Shoot")       // Command 
             ),
             new PathPoint(
-                new Translation2d(1.5,0),               // Position (meters)
-                new Rotation2d(Math.toRadians(180)),     // Rotation (rad)
-                2,    // Speed (m/s)
-                new PrintCommand("Intake")       // Command 
-            ),
-            new PathPoint(
-                new Translation2d(0.5,-1),               // Position (meters)
-                new Rotation2d(Math.toRadians(180)),     // Rotation (rad)
+                new Translation2d(2.838,7.054),               // Position (meters)
+                new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
                 1,    // Speed (m/s)
-                new PrintCommand("Transfer to next")       // Command 
-            ),
-            new PathPoint(
-                new Translation2d(1.5,-2),               // Position (meters)
-                new Rotation2d(Math.toRadians(180)),     // Rotation (rad)
-                0,    // Speed (m/s)
-                new PrintCommand("Intake")       // Command 
+                new PrintCommand("Shoot")       // Command 
             )
         );
     }
