+++ src/main/java/frc/robot/auto/paths/rightStart/threePiece/RightMidCenterMidRight.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.auto.paths;
+package frc.robot.auto.paths.rightStart.threePiece;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
@@ -14,14 +14,14 @@ import frc.robot.auto.pathing.pathObjects.PathPoint;
  * Example pure pursuit path, showing a basic
  * and readable implementation.
  */
-public class LeftNearLeft extends Path {
+public class RightMidCenterMidRight extends Path {
     /**
      * Drives the robot 1.5 meters along the x
      * axis (which should be forward, if the settings
      * are properly configured within the PathingConstants
      * class), at the robots maximum speed.
      */
-    public LeftNearLeft() {
+    public RightMidCenterMidRight() {
         /**
          * Initializes Path super with the given
          * set of points, the first point passed
@@ -29,14 +29,20 @@ public class LeftNearLeft extends Path {
          */
         super(            
             new PathPoint(
-                new Translation2d(0,7.054),               // Position (meters)
+                new Translation2d(0,1.622),               // Position (meters)
                 new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
-                1,    // Speed (m/s)
+                5,    // Speed (m/s)
                 new PrintCommand("Shoot")       // Command 
             ),
             new PathPoint(
-                new Translation2d(2.838,7.054),               // Position (meters)
+                new Translation2d(8.108,2.433),               // Position (meters)
                 new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
+                5,    // Speed (m/s)
+                new PrintCommand("Shoot")       // Command 
+            ),
+            new PathPoint(
+                new Translation2d(8.108,0.811),               // Position (meters)
+                new Rotation2d(Math.toRadians(Math.PI/2)),     // Rotation (rad)
                 1,    // Speed (m/s)
                 new PrintCommand("Shoot")       // Command 
             )
