+++ src/main/java/frc/robot/auto/paths/rightStart/threePiece/RightMidCenterRightMidCenter.java
@@ -2,11 +2,12 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.auto.paths;
+package frc.robot.auto.paths.rightStart.threePiece;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj2.command.PrintCommand;
+import frc.robot.Constants;
 import frc.robot.auto.pathing.pathObjects.Path;
 import frc.robot.auto.pathing.pathObjects.PathPoint;
 
@@ -29,7 +30,7 @@ public class RightMidCenterRightMidCenter extends Path {
          */
         super(            
             new PathPoint(
-                new Translation2d(0,1.622),               // Starting Position (meters)
+                Constants.Paths.START_RIGHT,               // Starting Position (meters)
                 new Rotation2d(Math.toRadians(0)),     // Rotation (rad)
                 1,    // Speed (m/s)
                 new PrintCommand("Started here")       // Command
